:bucket: refactor
